Sửa lại cấu trúc thư mục. Sort sử dụng enum, handle paging có size quá to.
