--- src/main/java/com/huy/mse/brand/Brand.java
@@ -1,33 +1,0 @@
-package com.huy.mse.brand;
-
-import com.huy.mse.product.Product;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDate;
-import java.util.List;
-
-@Entity
-@Table(name = "brand")
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Brand {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String name;
-
-    private String description;
-
-    @CreationTimestamp
-    @Column(name = "created_at")
-    private LocalDate createdAt;
-
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = "brand")
-    private List<Product> products;
-}
