--- src/main/java/com/huy/mse/brand/BrandDto.java
@@ -1,20 +1,0 @@
-package com.huy.mse.brand;
-
-import jakarta.validation.constraints.NotEmpty;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDate;
-
-@Getter
-@Setter
-public class BrandDto {
-    private long id;
-
-    @NotEmpty(message = "Name is required")
-    private String name;
-
-    private String description;
-
-    private LocalDate createdAt;
-}
