--- src/main/java/com/huy/mse/brand/BrandMapper.java
@@ -1,11 +1,0 @@
-package com.huy.mse.brand;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface BrandMapper {
-
-    BrandDto toDto(Brand brand);
-
-    Brand toEntity(BrandDto brandDto);
-}
