--- src/main/java/com/huy/mse/brand/BrandRepository.java
@@ -1,10 +1,0 @@
-package com.huy.mse.brand;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BrandRepository extends JpaRepository<Brand, Long> {
-
-    Brand findByName(String name);
-}
