--- src/main/java/com/huy/mse/category/Category.java
@@ -1,33 +1,0 @@
-package com.huy.mse.category;
-
-import com.huy.mse.product.Product;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDate;
-import java.util.List;
-
-@Entity
-@Table(name = "category")
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String name;
-
-    private String description;
-
-    @CreationTimestamp
-    @Column(name = "created_at")
-    private LocalDate createdAt;
-
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = "category")
-    private List<Product> products;
-}
