--- src/main/java/com/huy/mse/category/CategoryMapper.java
@@ -1,10 +1,0 @@
-package com.huy.mse.category;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface CategoryMapper {
-    CategoryDto toDto(Category category);
-
-    Category toEntity(CategoryDto categoryDto);
-}
