--- src/main/java/com/huy/mse/category/CategoryRepository.java
@@ -1,8 +1,0 @@
-package com.huy.mse.category;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long>{
-
-    Category findByName(String name);
-}
