--- src/main/java/com/huy/mse/category/CategoryService.java
@@ -1,11 +1,0 @@
-package com.huy.mse.category;
-
-import java.util.List;
-
-public interface CategoryService {
-    List<CategoryDto> getAllCategories();
-    CategoryDto getCategoryById(long id);
-    CategoryDto createCategory(CategoryDto categoryDto);
-    CategoryDto updateCategory(CategoryDto categoryDto);
-    void deleteCategory(long id);
-}
