--- src/main/java/com/huy/mse/common/exception/ValidationHandler.java
@@ -1,41 +1,0 @@
-package com.huy.mse.common.exception;
-
-import com.huy.mse.common.HttpResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-@ControllerAdvice
-public class ValidationHandler extends ResponseEntityExceptionHandler {
-
-    private final Logger logger = LoggerFactory.getLogger(getClass());
-
-    @Override
-    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatusCode status, WebRequest request) {
-        StringBuilder sb = new StringBuilder();
-        ex.getBindingResult().getFieldErrors().forEach(e -> sb.append(e.getDefaultMessage()).append("; "));
-
-        HttpResponse response = new HttpResponse(HttpStatus.BAD_REQUEST.value(), HttpStatus.BAD_REQUEST, HttpStatus.BAD_REQUEST.getReasonPhrase(), sb.toString());
-
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<HttpResponse> handleAllException(Exception e) {
-        logger.error(e.getMessage());
-        return createResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
-    }
-
-    private ResponseEntity<HttpResponse> createResponseEntity(HttpStatus status, String message) {
-        HttpResponse httpResponse = new HttpResponse(status.value(), status, status.getReasonPhrase(), message);
-        return new ResponseEntity<>(httpResponse, status);
-    }
-}
