+++ src/main/java/com/huy/mse/config/WebConfig.java
@@ -1,0 +1,14 @@
+package com.huy.mse.config;
+
+import com.huy.mse.enums.StringToProductSortTypeConverter;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+    @Override
+    public void addFormatters(FormatterRegistry registry) {
+        registry.addConverter(new StringToProductSortTypeConverter());
+    }
+}
