+++ src/main/java/com/huy/mse/dto/params/ProductParams.java
@@ -1,0 +1,16 @@
+package com.huy.mse.dto.params;
+
+import com.huy.mse.enums.ProductSortType;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ProductParams {
+    private String name;
+    private String brandNameList;
+    private String categoryNameList;
+    private int pageNumber = 0;
+    private int pageSize = 10;
+    private ProductSortType sortBy;
+}
