+++ src/main/java/com/huy/mse/mapper/BrandMapper.java
@@ -1,0 +1,13 @@
+package com.huy.mse.mapper;
+
+import com.huy.mse.dto.BrandDto;
+import com.huy.mse.entity.Brand;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface BrandMapper {
+
+    BrandDto toDto(Brand brand);
+
+    Brand toEntity(BrandDto brandDto);
+}
