--- src/main/java/com/huy/mse/product/Product.java
@@ -1,48 +1,0 @@
-package com.huy.mse.product;
-
-import com.huy.mse.brand.Brand;
-import com.huy.mse.category.Category;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDate;
-
-@Entity
-@Table(name = "product")
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Product {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String name;
-
-    private String description;
-
-    private double price;
-
-    private int stock;
-
-    private String batchNumber;
-
-    private LocalDate expiryDate;
-
-    @Column(name = "image_url")
-    private String imageUrl;
-
-    @CreationTimestamp
-    private LocalDate createdAt;
-
-    @ManyToOne
-    @JoinColumn(name = "category_id")
-    private Category category;
-
-    @ManyToOne
-    @JoinColumn(name = "brand_id")
-    private Brand brand;
-}
