--- src/main/java/com/huy/mse/product/ProductRepository.java
@@ -1,7 +1,0 @@
-package com.huy.mse.product;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-
-public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
-}
