+++ src/main/java/com/huy/mse/repository/BrandRepository.java
@@ -1,0 +1,11 @@
+package com.huy.mse.repository;
+
+import com.huy.mse.entity.Brand;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BrandRepository extends JpaRepository<Brand, Long> {
+
+    Brand findByName(String name);
+}
