+++ src/main/java/com/huy/mse/repository/ProductRepository.java
@@ -1,0 +1,8 @@
+package com.huy.mse.repository;
+
+import com.huy.mse.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
+}
