+++ src/main/java/com/huy/mse/service/CategoryService.java
@@ -1,0 +1,13 @@
+package com.huy.mse.service;
+
+import com.huy.mse.dto.CategoryDto;
+
+import java.util.List;
+
+public interface CategoryService {
+    List<CategoryDto> getAllCategories();
+    CategoryDto getCategoryById(long id);
+    CategoryDto createCategory(CategoryDto categoryDto);
+    CategoryDto updateCategory(CategoryDto categoryDto);
+    void deleteCategory(long id);
+}
