+++ src/main/java/com/huy/mse/service/impl/CategoryServiceImpl.java
@@ -1,0 +1,68 @@
+package com.huy.mse.service.impl;
+
+import com.huy.mse.dto.CategoryDto;
+import com.huy.mse.mapper.CategoryMapper;
+import com.huy.mse.repository.CategoryRepository;
+import com.huy.mse.entity.Category;
+import com.huy.mse.service.CategoryService;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class CategoryServiceImpl implements CategoryService {
+
+    private final CategoryRepository categoryRepository;
+    private final CategoryMapper categoryMapper;
+
+    public CategoryServiceImpl(CategoryRepository categoryRepository, CategoryMapper categoryMapper) {
+        this.categoryRepository = categoryRepository;
+        this.categoryMapper = categoryMapper;
+    }
+
+    public List<CategoryDto> getAllCategories() {
+        return categoryRepository.findAll()
+                .stream()
+                .map(categoryMapper::toDto)
+                .toList();
+    }
+
+    @Override
+    public CategoryDto getCategoryById(long id) {
+        return categoryRepository.findById(id)
+                .map(categoryMapper::toDto)
+                .orElseThrow(() -> new RuntimeException("Category not found"));
+    }
+
+    @Override
+    public CategoryDto createCategory(CategoryDto categoryDto) {
+        if (categoryRepository.findByName(categoryDto.getName()) != null) {
+            throw new RuntimeException("Category already exists");
+        }
+
+        Category category = categoryMapper.toEntity(categoryDto);
+        return categoryMapper.toDto(categoryRepository.save(category));
+    }
+
+    @Override
+    public CategoryDto updateCategory(CategoryDto categoryDto) {
+        Category category = categoryRepository.findById(categoryDto.getId())
+                .orElseThrow(() -> new RuntimeException("Category not found"));
+        category.setName(categoryDto.getName());
+        category.setDescription(categoryDto.getDescription());
+
+        return categoryMapper.toDto(categoryRepository.save(category));
+    }
+
+
+    @Override
+    public void deleteCategory(long id) {
+        Category category = categoryRepository.findById(id).orElse(null);
+
+        if (category == null) {
+            throw new RuntimeException("Category not found");
+        }
+
+        categoryRepository.delete(category);
+    }
+}
